Support /initialSync in synchrotron worker
